backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
anthropic>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
gitpython>=3.1.40
pypdf>=4.0.0
python-multipart>=0.0.9
slowapi>=0.1.9
python-dotenv>=1.0.0
httpx>=0.27.0
langchain-text-splitters>=0.2.0
